Become_author implemented. Bug fixed in crud.delete_refresh_token

+++ backend/API/crud.py
@@ -9,7 +9,6 @@ def get_user(db: Session, email: str):
 def get_scopes(db: Session, email: str):
     user = db.query(db_models.Users).filter(db_models.Users.email == email).first()
     scopes = []
-    print(user.author)
     if user.author != []:
         scopes += ['author']
     return scopes
@@ -25,6 +24,7 @@ def create_user(db: Session, user: schemes.UserCreate):
     db.refresh(db_user)
     return db_user
 
+
 def create_refresh_token(db: Session, refresh_token: schemes.RefreshToken):
     db_refresh_token = db_models.RefreshTokens(**refresh_token.dict())
     db.add(db_refresh_token)
@@ -37,5 +37,14 @@ def get_refresh_token(db: Session, refresh_token: uuid.UUID):
 
 def delete_refresh_token(db: Session, refresh_token: uuid.UUID):
     db_refresh_token = db.query(db_models.RefreshTokens).filter(db_models.RefreshTokens.uuid == refresh_token).first()
-    db.delete(db_refresh_token)
-    db.commit
+    if db_refresh_token:
+        db.delete(db_refresh_token)
+        db.commit()
+
+
+def create_author(db: Session, author: schemes.AuthorCreate):
+    db_author = db_models.Authors(user_id=author.user_id)
+    db.add(db_author)
+    db.commit()
+    db.refresh(db_author)
+    return db_author
